refactor: 폴더 구조 변경, 파일 나누기
