+++ src/components/MarkdownRenderer/StyledMarkdown.tsx
@@ -1,10 +1,6 @@
 import styled from "styled-components";
 import ReactMarkdown from "react-markdown";
-import remarkGfm from "remark-gfm";
-import remarkBreaks from "remark-breaks";
-import rehypeHighlight from "rehype-highlight";
 
-// ChatGPT 스타일 설정
 export const StyledMarkdown = styled(ReactMarkdown)`
   /* 본문 텍스트 스타일 */
   p {
@@ -118,21 +114,3 @@ export const StyledMarkdown = styled(ReactMarkdown)`
     font-style: italic;
   }
 `;
-
-const removeIndentation = (content: string) => {
-  return content.replace(/^[ \t]+/gm, ""); // 각 줄의 앞부분 공백 제거
-};
-
-const MarkdownStyles: React.FC<{ children: string }> = ({ children }) => {
-  const processedContent = removeIndentation(children);
-  return (
-    <StyledMarkdown
-      remarkPlugins={[remarkGfm, remarkBreaks]}
-      rehypePlugins={[rehypeHighlight]}
-    >
-      {processedContent}
-    </StyledMarkdown>
-  );
-};
-
-export default MarkdownStyles;
