+++ src/widgets/Blog/posts/1/241010_input_output.ts
