+++ src/widgets/Blog/posts/1/241010_int_char_bool_string.ts
