+++ src/widgets/Blog/posts/1/241014_vector.ts
