+++ src/widgets/Blog/posts/generate-posts-index.js
