+++ src/widgets/Blog/posts/index.ts
