+++ src/widgets/Board/BoardTypes.ts
