+++ src/widgets/Intro/IntroData.ts
@@ -1,0 +1,20 @@
+export const TITLE = {
+  name: "송강산",
+  title: "빠져드는 개발자",
+};
+
+export const INTRODUCTION = `
+    -   문제 해결에 깊이 **몰입**하고,
+        끊임없이 **탐구**하는 개발자입니다.
+    -   **사용자 경험**을 최우선으로, 
+        기술을 유익하고 편리하게 
+        설계하려고 **노력**합니다.
+    -   팀원들과 협업, 지식 공유에 있어서 
+        적극적이고 능숙하게 **소통**합니다. 
+`;
+
+export const SKILLS = [
+  { label: "프론트엔드", level: 3 },
+  { label: "백엔드", level: 2 },
+  { label: "임베디드", level: 2 },
+];
