+++ src/widgets/Title/TitleStyles.ts
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const TitleContainer = styled.div`
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  margin: 0;
+  padding: 0;
+`;
