+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/main.tsx","./src/vite-env.d.ts","./src/app/app.tsx","./src/app/index.tsx","./src/features/card/index.tsx","./src/features/card/ui.tsx","./src/features/openboardbutton/index.tsx","./src/features/openboardbutton/ui.tsx","./src/pages/home/home.tsx","./src/pages/home/index.tsx","./src/shared/constants/boardtypes.tsx","./src/shared/constants/colors.tsx","./src/shared/posts/000000_contents.ts","./src/shared/posts/241010_input_output.ts","./src/shared/posts/241010_int_char_bool_string.ts","./src/shared/posts/241014_queue_priority_queue.ts","./src/shared/posts/241014_vector.ts","./src/shared/posts/index.ts","./src/shared/styles/markdownstyles.tsx","./src/widgets/index.tsx","./src/widgets/blog/index.tsx","./src/widgets/blog/ui.tsx","./src/widgets/board/index.tsx","./src/widgets/board/ui.tsx","./src/widgets/intro/index.tsx","./src/widgets/intro/ui.tsx","./src/widgets/title/index.tsx","./src/widgets/title/ui.tsx","./src/widgets/title/model/usemodel.tsx"],"version":"5.6.3"}
+{"root":["./src/main.tsx","./src/vite-env.d.ts","./src/app/app.tsx","./src/app/index.tsx","./src/components/markdownrenderer/styledmarkdown.tsx","./src/components/markdownrenderer/index.tsx","./src/components/modal/modalstyles.tsx","./src/components/modal/index.tsx","./src/entities/lottie/index.tsx","./src/entities/model/index.ts","./src/features/card/cardstyles.ts","./src/features/card/index.tsx","./src/features/openboardbutton/index.tsx","./src/features/spiral/spiralstyles.ts","./src/features/spiral/index.tsx","./src/features/spiral/ui.tsx","./src/features/titleheader/index.tsx","./src/pages/home/homestyles.tsx","./src/pages/home/index.tsx","./src/shared/constants/colors.tsx","./src/widgets/index.tsx","./src/widgets/blog/blogtitles.ts","./src/widgets/blog/index.tsx","./src/widgets/blog/posts/index.ts","./src/widgets/blog/posts/1/241010_input_output.ts","./src/widgets/blog/posts/1/241010_int_char_bool_string.ts","./src/widgets/blog/posts/1/241014_queue_priority_queue.ts","./src/widgets/blog/posts/1/241014_vector.ts","./src/widgets/board/boardstyles.tsx","./src/widgets/board/boardtypes.ts","./src/widgets/board/index.tsx","./src/widgets/intro/introdata.ts","./src/widgets/intro/introstyles.ts","./src/widgets/intro/index.tsx","./src/widgets/title/model.tsx","./src/widgets/title/titlestyles.ts","./src/widgets/title/index.tsx"],"version":"5.6.3"}
